new commit

+++ resources/views/roles/index.blade.php
@@ -8,9 +8,9 @@
                 <h2>Role Management</h2>
             </div>
             <div class="pull-right">
-                {{-- @can('role-create') --}}
+                @can('role-create')
                     <a class="btn btn-success" href="{{ route('roles.create') }}"> Create New Role</a>
-                {{-- @endcan --}}
+                @endcan
             </div>
         </div>
     </div>
@@ -29,21 +29,21 @@
             <th>Name</th>
             <th width="280px">Action</th>
         </tr>
-        
+
         @foreach ($roles as $key => $role)
             <tr>
                 <td>{{ ++$i }}</td>
                 <td>{{ $role->name }}</td>
                 <td>
                     <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Show</a>
-                    {{-- @can('role-edit')
+                    @can('role-edit')
                         <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                    @endcan --}}
-                    {{-- @can('role-delete')
+                    @endcan
+                    @can('role-delete')
                         {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                         {!! Form::close() !!}
-                    @endcan --}}
+                    @endcan
                 </td>
             </tr>
         @endforeach
